% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana_function.R
\name{banana}
\alias{banana}
\title{A function for missed calls on bananaphone}
\usage{
banana(answer, rings, caller)
}
\arguments{
\item{answer}{would you like to answer the call}

\item{rings}{how many times did bananaphone ring}

\item{caller}{who called}
}
\description{
A function for missed calls on bananaphone
}
\examples{
banana("Yes", 2, "Ruth")
}
